+++ Detectors/Upgrades/ALICE3/TRK/base/include/TRKBase/GeometryTGeo.h
@@ -235,8 +235,7 @@ class GeometryTGeo : public o2::detectors::DetMatrixCache
   std::vector<float> mCacheRefXMLOT;     /// cache for X of ML and OT
   std::vector<float> mCacheRefAlphaMLOT; /// cache for sensor ref alpha ML and OT
 
-  eLayout mLayoutML; // Type of segmentation for the middle layers
-  eLayout mLayoutOT; // Type of segmentation for the outer layers
+  eMLOTLayout mLayoutMLOT; // ML and OT detector layout design
 
  private:
   static std::unique_ptr<o2::trk::GeometryTGeo> sInstance;
